move fsChunks into it own proposal file, https://github.com/microsoft/vscode/issues/131165
