+++ src/vs/workbench/api/common/extHost.api.impl.ts
@@ -914,7 +914,7 @@ export function createApiFactoryAndRegisterActors(accessor: ServicesAccessor): I
 				return extHostTask.registerTaskProvider(extension, type, provider);
 			},
 			registerFileSystemProvider(scheme, provider, options) {
-				return extHostFileSystem.registerFileSystemProvider(extension.identifier, scheme, provider, options);
+				return extHostFileSystem.registerFileSystemProvider(extension, scheme, provider, options);
 			},
 			get fs() {
 				return extHostConsumerFileSystem.value;
