+++ src/vs/workbench/services/extensions/common/extensionsApiProposals.ts
@@ -7,6 +7,7 @@
 
 export const allApiProposals = Object.freeze({
 	editorInsets: 'https://raw.githubusercontent.com/microsoft/vscode/main/src/vscode-dts/vscode.proposed.editorInsets.d.ts',
+	fsChunks: 'https://raw.githubusercontent.com/microsoft/vscode/main/src/vscode-dts/vscode.proposed.fsChunks.d.ts',
 	languageStatus: 'https://raw.githubusercontent.com/microsoft/vscode/main/src/vscode-dts/vscode.proposed.languageStatus.d.ts'
 });
 export type ApiProposalName = keyof typeof allApiProposals;
