+++ src/vscode-dts/vscode.proposed.d.ts
@@ -267,19 +267,6 @@ declare module 'vscode' {
 
 	//#endregion
 
-
-
-	//#region fsChunks: https://github.com/microsoft/vscode/issues/84515
-
-	export interface FileSystemProvider {
-		open?(resource: Uri, options: { create: boolean; }): number | Thenable<number>;
-		close?(fd: number): void | Thenable<void>;
-		read?(fd: number, pos: number, data: Uint8Array, offset: number, length: number): number | Thenable<number>;
-		write?(fd: number, pos: number, data: Uint8Array, offset: number, length: number): number | Thenable<number>;
-	}
-
-	//#endregion
-
 	//#region textSearchProvider: https://github.com/microsoft/vscode/issues/59921
 
 	/**
